test(etl): テストファイル構成を実装ファイルに合わせて修正
